+++ test/groovy/org/ods/services/OpenShiftServiceSpec.groovy
@@ -102,12 +102,13 @@ class OpenShiftServiceSpec extends SpecHelper {
 
         then:
         1 * steps.sh(
-            script: 'helm -n foo secrets diff upgrade --install --force -f values.yml -f values-dev.yml --set imageTag=6f8db5fb --no-color bar ./',
+            script: 'HELM_DIFF_IGNORE_UNKNOWN_FLAGS=true helm -n foo secrets diff upgrade --install --atomic --force -f values.yml -f values-dev.yml --set imageTag=6f8db5fb --no-color --three-way-merge --normalize-manifests bar ./',
             label: 'Show diff explaining what helm upgrade would change for release bar in foo'
         )
         1 * steps.sh(
             script: 'helm -n foo secrets upgrade --install --atomic --force -f values.yml -f values-dev.yml --set imageTag=6f8db5fb bar ./',
-            label: 'Upgrade Helm release bar in foo'
+            label: 'Upgrade Helm release bar in foo',
+            returnStatus: true
         )
     }